// src/models/supplierModel.ts
export interface SupplierModel {
    photo_url: string;
    name: string;
    slug: string;
    product: string;
    email: string;
    active: number;
    categories: string[];
    price: number;
    contact: string;
    isTasking: number;
    createdAt: string;
    updatedAt: string;
    isDeleted: boolean;
    _id: string;
}